Optional to save cache to file

+++ api/modules/cache/index.ts
@@ -2,21 +2,36 @@ import { ChatMessage } from "bing-chat";
 import { existsSync, mkdirSync, readFileSync, writeFileSync } from "fs";
 
 class Cache {
+  private cache: {
+    [index: string]: ChatMessage;
+  } = {};
   private cacheFile = "./cache/cache.json";
+  private isFileCache = process.env.FILE_CACHE || false;
+
   constructor() {
-    if (!existsSync("./cache")) mkdirSync("./cache");
-    if (!existsSync(this.cacheFile)) writeFileSync(this.cacheFile, "{}");
+    if (this.isFileCache) {
+      if (!existsSync("./cache")) mkdirSync("./cache");
+      if (!existsSync(this.cacheFile)) writeFileSync(this.cacheFile, "{}");
+    }
   }
 
   getData(userId: number) {
-    return JSON.parse(readFileSync(this.cacheFile).toString())[userId] as ChatMessage;
+    if (this.isFileCache) {
+      return JSON.parse(readFileSync(this.cacheFile).toString())[userId] as ChatMessage;
+    } else {
+      return this.cache[userId] as ChatMessage;
+    }
   }
 
   setData(userId: number, data: ChatMessage) {
-    let cacheData = JSON.parse(readFileSync(this.cacheFile).toString());
-    cacheData[userId] = data;
+    if (this.isFileCache) {
+      let cacheData = JSON.parse(readFileSync(this.cacheFile).toString());
+      cacheData[userId] = data;
 
-    writeFileSync(this.cacheFile, JSON.stringify(cacheData, null, 2));
+      writeFileSync(this.cacheFile, JSON.stringify(cacheData, null, 2));
+    } else {
+      this.cache[userId] = data;
+    }
   }
 }
